+++ cli/src/it/scala/co/topl/brambl/cli/CommonFunctions.scala
@@ -40,7 +40,7 @@ trait CommonFunctions extends PolicyTemplates {
           _ <- IO.sleep(5.seconds)
         } yield queryRes)
           .iterateUntil(_ == ExitCode.Success),
-        60.seconds
+        240.seconds
       )
       ALICE_TO_ADDRESS <- walletController(ALICE_WALLET).currentaddress("self", "default", None)
       _ <- IO.println(s"Alice's address is $ALICE_TO_ADDRESS")
@@ -82,7 +82,7 @@ trait CommonFunctions extends PolicyTemplates {
           _ <- IO.sleep(5.seconds)
         } yield queryRes)
           .iterateUntil(_ == ExitCode.Success),
-        60.seconds
+        240.seconds
       )
     } yield res
   }